pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "assemblyai>=0.42.1",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.29.0",
    "google-generativeai==0.8.3",
    "gunicorn>=23.0.0",
    "murf>=2.0.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "speechrecognition>=3.14.3",
]
